gasket: Add downstream driver for Coral AI hardware
The in-tree staging driver is insufficient, Google only updated their
downstream code. That resides in a linux-imx repo which is apparently
holding BSP code for the dev boards as well. Unfortunately, it's hosted
on a git server that does not permit stable tarball downloads.

Therefore, we take a snapshot of the few files needed to avoid pulling
a complete kernel tree only for those. The actual module recipe is
trivial then.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>