Limit apt snapshot traffic
This does not limit the speed of debootstrap, however, at least we have
some method to limit the speed of the apt to snapshot.d.o.

Another possible way is to use the tc to limit the bandwidth of the
whole interface, but it's more brutal, and cannot handle the ip change
of the snapshot.d.o.

A more elegant way should be figured out to solve this issue. This one
should be only a not-so-good workaround.

Signed-off-by: Su Bao Cheng <baocheng.su@siemens.com>